Updates for use with OSX

+++ Annex/NSString+Annex.m
@@ -30,7 +30,7 @@
 {
     const char *cStr = [string UTF8String];
 	unsigned char digest[16];
-	CC_MD5( cStr, strlen(cStr), digest);
+	CC_MD5( cStr, (unsigned int)strlen(cStr), digest);
     
 	NSMutableString *output = [NSMutableString stringWithCapacity:CC_MD5_DIGEST_LENGTH * 2];
     
@@ -47,7 +47,7 @@
     
 	uint8_t digest[CC_SHA1_DIGEST_LENGTH];
     
-	CC_SHA1(data.bytes, data.length, digest);
+	CC_SHA1(data.bytes, (unsigned int)data.length, digest);
     
 	NSMutableString* output = [NSMutableString stringWithCapacity:CC_SHA1_DIGEST_LENGTH * 2];
